Don't compare nlmsg_pid to the process pid.
The nlmsg_pid is opaque PORT ID that doesn't have to be equal to the process id. In this case it is assigned by kernel and only equals to process id if this is the first netlink socket. When another netlink socket already exists, this one gets arbitrary value assigned and then all received messages are skipped.

One could retrieve the assigned PORT ID and test against it, but there doesn't seem to be reason to test it at all.